backend/FAST-IBAN_Project/requirements.txt
basemap==1.3.8
basemap_data==1.3.2
CairoSVG==2.7.1
Cartopy==0.22.0
imageio==2.34.1
matplotlib==3.7.2
netCDF4==1.6.5
numpy==1.25.2
pandas==2.0.3
PyYAML==6.0.1
scipy==1.13.0
xarray==2024.3.0
